// app/types.ts
export type Id = string | number;

export type Column = {
  id: Id;
  title: string;
  description:string;
  colums:[{
    id: Id;
    title: string;
    decription:string;
  },{
    id: Id;
    title: string;
    decription:string;
  },{
    id: Id;
    title: string;
    decription:string;
  }]
};

export type Task = {
  id: Id;
  columnId: Id;
  content: string;
};